requirements.txt
amqp==5.1.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.6
djangorestframework==3.13.1
kombu==5.2.4
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
